Take a stance on for loops: use const references for each item.
This is safer, because the code isn't at the mercy of the compiler for what
destructors run on the for loop elements (e.g., std::string). This was bad
when we would store a const char * reference to a std::string's data(), only
for the std::string to go out of scope at the end of the for loop and the
const char * data to be freed.