Use exceptions to alleviate error handling pains.
This goes against my preference, but using exceptions really does knock out
all of the pains we had with errors. This will be a good exercise for the
procedural reader to see if this works. :)

One general worry I have is whether this will lead to confusing, minimal error
messages on the command line. Perhaps we will need to add extra exception
handling at various points in the code to get around this issue.