ci: actually, do build btool, but use circleci image.
The btool container was doubling for a dev and ci container. It was huge, and took
to long to pull, build, and push for circleci. Let's use a circleci image to speed
up ci and use btool as an example image to try out btool. I moved the rest of the
dev stuff to andrew-workstation/image/dev.