app::collector::registry: try to catch flakey http registry unmarshaling.
There is some weird issue where the YamlSerializer fails because either
the Index or Registry converter can't find the 'files' or 'nodes' key,
respectively. I wonder if it is some sort of buffer underflow from the
fstream. Let's add a test here to try to catch it, and also add a sync
to see if that fixes it.