Inject -lpthread into build via linker flags.
This is janky. We need to do it before linux builds don't automatically
link with libpthread, whereas macos builds do. This also won't make CI
pass, because the version of btool that we are building with does not
have this -lpthread flag in it.

Sigh. Building a tool that builds itself is more complicated than I
originally expected.